requirements.txt
langchain-ollama>=0.0.6
langchain-openai>=0.0.2
langchain-core>=0.1.20
python-dotenv>=1.0.0